[export] [call-v3] Begin migration of channel_init to templates.
Eventually for call-v3 we're going to want to have registration of filters generate the appropriate glue into the channel runtime to execute a call.

Begin that process now and gradually by introducing the new syntax and allowing a piecemeal migration to it - by the time we're done converting filters to the v3 APIs we'll also have the registration piece done.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/591614473](http://cl/591614473)

PiperOrigin-RevId: 591614473