[export] [Python] completion queue loop: only invoke Python’s signal handling in main thread
When invoked from a non-main thread, [`PyErr_CheckSignals`](https://docs.python.org/3/c-api/exceptions.html#c.PyErr_CheckSignals) call is a no-op, but we exiting nogil to execute it.

This change improves the performance by not performing GIL release/reacquire for signal handling unless we're in the main thread.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/758901710](http://cl/758901710)

PiperOrigin-RevId: 758901710