[export] [chaotic-good] External connector support
Add external connector support to chaotic good

This is honestly a bit of a mess, and we need to make it simpler just as soon as we've got cycles - there should be no need for chaotic good to be aware of posix vs non-posix in the server code.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/767430169](http://cl/767430169)

PiperOrigin-RevId: 767430169