[export] Create a standard config `:grpc_small_clients` for binary-size sensitive clients. This includes Android, iOS, and a newly-introduced opt-in flag --//:small_clients.
My assumption is that it will be rare for clients to take opt-out features
piecemeal - if they are size constrained, they will want every possible
opt-out. I've left the existing opt-outs in place for current users, but
future features can decide if they need their own flag or not.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/788506202](http://cl/788506202)

PiperOrigin-RevId: 788506202